Split sol-util to each platform
Riot don't have a filesystem also Contiki don't have, so spliting
sol-util to fix build.
